feat(FN-1745): move selected reported fees view model mapper to helper file
