+++ trade-finance-manager-ui/server/controllers/utilisation-reports/add-payment/index.ts
@@ -1,7 +1,7 @@
 import { Response } from 'express';
 import { getFormattedCurrencyAndAmount, getFormattedReportPeriodWithLongMonth } from '@ukef/dtfs2-common';
 import { format, parseISO } from 'date-fns';
-import { AddPaymentViewModel, RecordedPaymentDetailsViewModel, SelectedReportedFeesDetailsViewModel } from '../../../types/view-models';
+import { AddPaymentViewModel, RecordedPaymentDetailsViewModel } from '../../../types/view-models';
 import api from '../../../api';
 import { asUserSession } from '../../../helpers/express-session';
 import {
@@ -13,13 +13,13 @@ import { CustomExpressRequest } from '../../../types/custom-express-request';
 import { ValidatedAddPaymentFormValues } from '../../../types/add-payment-form-values';
 import { PRIMARY_NAVIGATION_KEYS } from '../../../constants';
 import {
-  getKeyToCurrencyAndAmountSortValueMap,
   parseValidatedAddPaymentFormValues,
   AddPaymentFormRequestBody,
   extractAddPaymentFormValuesAndValidateIfPresent,
   EMPTY_ADD_PAYMENT_FORM_VALUES,
+  mapToSelectedReportedFeesDetailsViewModel,
 } from '../helpers';
-import { SelectedFeeRecordsDetailsResponseBody, SelectedFeeRecordsPaymentDetailsResponse } from '../../../api-response-types';
+import { SelectedFeeRecordsPaymentDetailsResponse } from '../../../api-response-types';
 
 export type AddPaymentRequest = CustomExpressRequest<{
   reqBody: AddPaymentFormRequestBody;
@@ -32,28 +32,6 @@ const mapToRecordedPaymentDetailsViewModel = (payment: SelectedFeeRecordsPayment
     reference: payment.reference,
     formattedCurrencyAndAmount: getFormattedCurrencyAndAmount({ currency: payment.currency, amount: payment.amount }),
     formattedDateReceived: format(parseISO(payment.dateReceived), 'd MMM yyyy'),
-  };
-};
-
-const mapToSelectedReportedFeesDetailsViewModel = (selectedFeeRecordData: SelectedFeeRecordsDetailsResponseBody): SelectedReportedFeesDetailsViewModel => {
-  const reportedFeeDataSortValueMap = getKeyToCurrencyAndAmountSortValueMap(
-    selectedFeeRecordData.feeRecords.map((record) => ({ ...record.reportedFee, key: record.id })),
-  );
-  const reportedPaymentsDataSortValueMap = getKeyToCurrencyAndAmountSortValueMap(
-    selectedFeeRecordData.feeRecords.map((record) => ({ ...record.reportedPayments, key: record.id })),
-  );
-  return {
-    totalReportedPayments: getFormattedCurrencyAndAmount(selectedFeeRecordData.totalReportedPayments),
-    feeRecords: selectedFeeRecordData.feeRecords.map((record) => ({
-      id: record.id,
-      facilityId: record.facilityId,
-      exporter: record.exporter,
-      reportedFees: { formattedCurrencyAndAmount: getFormattedCurrencyAndAmount(record.reportedFee), dataSortValue: reportedFeeDataSortValueMap[record.id] },
-      reportedPayments: {
-        formattedCurrencyAndAmount: getFormattedCurrencyAndAmount(record.reportedPayments),
-        dataSortValue: reportedPaymentsDataSortValueMap[record.id],
-      },
-    })),
   };
 };
